@@ -6,14 +6,14 @@ homepage: https://github.com/awslabs/aws-lambda-dart-runtime
 documentation: https://awslabs.github.io/aws-lambda-dart-runtime/
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  json_annotation: ^3.0.0
+  json_annotation: ^4.0.1
 
 dev_dependencies:
-  json_serializable: ^3.2.3
+  json_serializable: ^4.1.0
   build_runner: ^1.7.2
-  test: ^1.9.4
-  mockito: ^4.1.1
-  pedantic: ^1.0.0
+  test: ^1.16.8
+  mockito: ^5.0.3
+  pedantic: ^1.11.0
