@@ -13,15 +13,11 @@ AwsCognitoEvent _$AwsCognitoEventFromJson(Map<String, dynamic> json) {
     region: json['region'] as String,
     userPoolId: json['userPoolId'] as String,
     userName: json['userName'] as String,
-    callerContext: (json['callerContext'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
-    request: json['request'] == null
-        ? null
-        : AwsCognitoRequest.fromJson(json['request'] as Map<String, dynamic>),
-    response: json['response'] == null
-        ? null
-        : AwsCognitoResponse.fromJson(json['response'] as Map<String, dynamic>),
+    callerContext: Map<String, String>.from(json['callerContext'] as Map),
+    request:
+        AwsCognitoRequest.fromJson(json['request'] as Map<String, dynamic>),
+    response:
+        AwsCognitoResponse.fromJson(json['response'] as Map<String, dynamic>),
   );
 }
 
@@ -40,15 +36,17 @@ Map<String, dynamic> _$AwsCognitoEventToJson(AwsCognitoEvent instance) =>
 AwsCognitoRequest _$AwsCognitoRequestFromJson(Map<String, dynamic> json) {
   return AwsCognitoRequest(
     userAttributes: json['userAttributes'] as Map<String, dynamic>,
-    validationData: (json['validationData'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
-    clientMetadata: (json['clientMetadata'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
-    newDeviceUsed: json['newDeviceUsed'] as bool,
+    validationData: (json['validationData'] as Map<String, dynamic>?)?.map(
+          (k, e) => MapEntry(k, e as String),
+        ) ??
+        {},
+    clientMetadata: (json['clientMetadata'] as Map<String, dynamic>?)?.map(
+          (k, e) => MapEntry(k, e as String),
+        ) ??
+        {},
+    newDeviceUsed: json['newDeviceUsed'] as bool? ?? false,
     codeParameter: json['codeParameter'] as String,
-    password: json['password'] as String,
+    password: json['password'] as String?,
     groupConfiguration: json['groupConfiguration'] == null
         ? null
         : AwsGroupConfiguration.fromJson(
@@ -69,25 +67,27 @@ Map<String, dynamic> _$AwsCognitoRequestToJson(AwsCognitoRequest instance) =>
 
 AwsCognitoResponse _$AwsCognitoResponseFromJson(Map<String, dynamic> json) {
   return AwsCognitoResponse(
-    autoConfirmUser: json['autoConfirmUser'] as bool,
-    autoVerifyEmail: json['autoVerifyEmail'] as bool,
-    autoVerifyPhone: json['autoVerifyPhone'] as bool,
+    autoConfirmUser: json['autoConfirmUser'] as bool? ?? false,
+    autoVerifyEmail: json['autoVerifyEmail'] as bool? ?? false,
+    autoVerifyPhone: json['autoVerifyPhone'] as bool? ?? false,
     claimsOverrideDetails: json['claimsOverrideDetails'] == null
         ? null
         : AwsClaimOverrideDetails.fromJson(
             json['claimsOverrideDetails'] as Map<String, dynamic>),
-    userAttributes: (json['userAttributes'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
-    finalUserStatus: json['finalUserStatus'] as String,
-    desiredDeliveryMediums: (json['desiredDeliveryMediums'] as List)
-        ?.map((e) => e as String)
-        ?.toList(),
-    forceAliasCreation: json['forceAliasCreation'] as bool,
-    messageAction: json['messageAction'] as String,
-    smsMessage: json['smsMessage'] as String,
-    emailMessage: json['emailMessage'] as String,
-    emailSubject: json['emailSubject'] as String,
+    userAttributes: (json['userAttributes'] as Map<String, dynamic>?)?.map(
+          (k, e) => MapEntry(k, e as String),
+        ) ??
+        {},
+    finalUserStatus: json['finalUserStatus'] as String? ?? '',
+    desiredDeliveryMediums: (json['desiredDeliveryMediums'] as List<dynamic>?)
+            ?.map((e) => e as String)
+            .toList() ??
+        [],
+    forceAliasCreation: json['forceAliasCreation'] as bool? ?? false,
+    messageAction: json['messageAction'] as String? ?? '',
+    smsMessage: json['smsMessage'] as String? ?? '',
+    emailMessage: json['emailMessage'] as String? ?? '',
+    emailSubject: json['emailSubject'] as String? ?? '',
   );
 }
 
@@ -110,14 +110,14 @@ Map<String, dynamic> _$AwsCognitoResponseToJson(AwsCognitoResponse instance) =>
 AwsGroupConfiguration _$AwsGroupConfigurationFromJson(
     Map<String, dynamic> json) {
   return AwsGroupConfiguration(
-    groupsToOverride:
-        (json['groupsToOverride'] as List)?.map((e) => e as String)?.toList(),
-    iamRolesToOverride:
-        (json['iamRolesToOverride'] as List)?.map((e) => e as String)?.toList(),
+    groupsToOverride: (json['groupsToOverride'] as List<dynamic>)
+        .map((e) => e as String)
+        .toList(),
+    iamRolesToOverride: (json['iamRolesToOverride'] as List<dynamic>)
+        .map((e) => e as String)
+        .toList(),
     preferredRole: json['preferredRole'] as String,
-    clientMetadata: (json['clientMetadata'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
+    clientMetadata: Map<String, String>.from(json['clientMetadata'] as Map),
   );
 }
 
@@ -134,15 +134,12 @@ AwsClaimOverrideDetails _$AwsClaimOverrideDetailsFromJson(
     Map<String, dynamic> json) {
   return AwsClaimOverrideDetails(
     claimsToAddOrOverride:
-        (json['claimsToAddOrOverride'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
-    claimsToSuppress:
-        (json['claimsToSuppress'] as List)?.map((e) => e as String)?.toList(),
-    groupOverrideDetails: json['groupOverrideDetails'] == null
-        ? null
-        : AwsGroupConfiguration.fromJson(
-            json['groupOverrideDetails'] as Map<String, dynamic>),
+        Map<String, String>.from(json['claimsToAddOrOverride'] as Map),
+    claimsToSuppress: (json['claimsToSuppress'] as List<dynamic>)
+        .map((e) => e as String)
+        .toList(),
+    groupOverrideDetails: AwsGroupConfiguration.fromJson(
+        json['groupOverrideDetails'] as Map<String, dynamic>),
   );
 }
 
