@@ -8,11 +8,9 @@ part of 's3_event.dart';
 
 AwsS3Event _$AwsS3EventFromJson(Map<String, dynamic> json) {
   return AwsS3Event(
-    records: (json['Records'] as List)
-        ?.map((e) => e == null
-            ? null
-            : AwsS3EventRecord.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
+    records: (json['Records'] as List<dynamic>)
+        .map((e) => AwsS3EventRecord.fromJson(e as Map<String, dynamic>))
+        .toList(),
   );
 }
 
@@ -26,23 +24,14 @@ AwsS3EventRecord _$AwsS3EventRecordFromJson(Map<String, dynamic> json) {
     eventName: json['eventName'] as String,
     eventSource: json['eventSource'] as String,
     awsRegion: json['awsRegion'] as String,
-    eventTime: json['eventTime'] == null
-        ? null
-        : DateTime.parse(json['eventTime'] as String),
+    eventTime: DateTime.parse(json['eventTime'] as String),
     eventVersion: json['eventVersion'] as String,
-    userIdentity: json['userIdentity'] == null
-        ? null
-        : AwsS3UserIdentity.fromJson(
-            json['userIdentity'] as Map<String, dynamic>),
-    requestParameters: (json['requestParameters'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
-    responseElements: (json['responseElements'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
-    s3: json['s3'] == null
-        ? null
-        : AwsS3Data.fromJson(json['s3'] as Map<String, dynamic>),
+    userIdentity: AwsS3UserIdentity.fromJson(
+        json['userIdentity'] as Map<String, dynamic>),
+    requestParameters:
+        Map<String, String>.from(json['requestParameters'] as Map),
+    responseElements: Map<String, String>.from(json['responseElements'] as Map),
+    s3: AwsS3Data.fromJson(json['s3'] as Map<String, dynamic>),
   );
 }
 
@@ -51,7 +40,7 @@ Map<String, dynamic> _$AwsS3EventRecordToJson(AwsS3EventRecord instance) =>
       'eventVersion': instance.eventVersion,
       'eventSource': instance.eventSource,
       'awsRegion': instance.awsRegion,
-      'eventTime': instance.eventTime?.toIso8601String(),
+      'eventTime': instance.eventTime.toIso8601String(),
       'eventName': instance.eventName,
       'userIdentity': instance.userIdentity,
       'requestParameters': instance.requestParameters,
@@ -63,12 +52,8 @@ AwsS3Data _$AwsS3DataFromJson(Map<String, dynamic> json) {
   return AwsS3Data(
     s3SchemaVersion: json['s3SchemaVersion'] as String,
     configurationId: json['configurationId'] as String,
-    bucket: json['bucket'] == null
-        ? null
-        : AwsS3Bucket.fromJson(json['bucket'] as Map<String, dynamic>),
-    object: json['object'] == null
-        ? null
-        : AWSS3EventObject.fromJson(json['object'] as Map<String, dynamic>),
+    bucket: AwsS3Bucket.fromJson(json['bucket'] as Map<String, dynamic>),
+    object: AWSS3EventObject.fromJson(json['object'] as Map<String, dynamic>),
   );
 }
 
@@ -82,10 +67,8 @@ Map<String, dynamic> _$AwsS3DataToJson(AwsS3Data instance) => <String, dynamic>{
 AwsS3Bucket _$AwsS3BucketFromJson(Map<String, dynamic> json) {
   return AwsS3Bucket(
     name: json['name'] as String,
-    ownerIdentity: json['ownerIdentity'] == null
-        ? null
-        : AwsS3UserIdentity.fromJson(
-            json['ownerIdentity'] as Map<String, dynamic>),
+    ownerIdentity: AwsS3UserIdentity.fromJson(
+        json['ownerIdentity'] as Map<String, dynamic>),
     arn: json['arn'] as String,
   );
 }
